requirements.txt
flask
requests
psycopg2-binary
bs4
schedule
datetime